Post usuário arrumado e otimização no código
