+++ src/enponits/getAllDocente.ts
@@ -1,0 +1,17 @@
+import { Request, Response } from 'express'
+import { DocenteDataBase } from '../data/DocenteDatabase'
+
+export const getAllDocente = async (req: Request, res: Response): Promise <void > => {
+
+    try {
+        const docenteDB = new DocenteDataBase()
+        const result = await docenteDB.getDocentes()
+
+        res.status(200).send(result)
+    }
+
+    catch (error: any) {
+        res.status(400).send(error.sqlMessage || error.message)
+    }
+
+}
